display: warning when moulitest is not installed

+++ includes/get_next_line.sh
@@ -198,25 +198,30 @@ function check_gnl_macro
 function check_gnl_moulitest
 {
 	local RET0 TOTAL
-	rm -f "$RETURNPATH"/.mymoulitest
-	cd "$RETURNPATH/moulitest/get_next_line_tests/"
-	make 1> "$RETURNPATH"/.mymoulitest 2>&1
-	cd "$RETURNPATH"
-	RET0=`cat .mymoulitest | sed 's/\^\[\[[0-9;]*m//g' | sed 's/\^\[\[0m//g' | sed 's/\$$//' | grep "END OF UNIT TESTS"`
-	if [ "$RET0" == "" ]
+	if [ -d moulitest ]
 	then
-		printf $C_RED"  Fatal error: moulitest cannot compile (see details)"$C_CLEAR
-	else
-		RET0=`cat -e .mymoulitest | grep FAIL | sed 's/\^\[\[[0-9;]*m//g' | sed 's/\^\[\[0m//g' | sed 's/\$$//' | awk 'BEGIN {OFS = ""} {print "  ",$0}'`
-		if [ "$RET0" != "" ]
+		rm -f "$RETURNPATH"/.mymoulitest
+		cd "$RETURNPATH/moulitest/get_next_line_tests/"
+		make 1> "$RETURNPATH"/.mymoulitest 2>&1
+		cd "$RETURNPATH"
+		RET0=`cat .mymoulitest | sed 's/\^\[\[[0-9;]*m//g' | sed 's/\^\[\[0m//g' | sed 's/\$$//' | grep "END OF UNIT TESTS"`
+		if [ "$RET0" == "" ]
 		then
-			TOTAL=`printf "$RET0" | wc -l | sed 's/ //g'`
-			printf $C_RED"  $TOTAL failed test(s)"$C_CLEAR
+			printf $C_RED"  Fatal error: moulitest cannot compile (see details)"$C_CLEAR
 		else
-			printf $C_GREEN"  All Unit Tests passed"$C_CLEAR
+			RET0=`cat -e .mymoulitest | grep FAIL | sed 's/\^\[\[[0-9;]*m//g' | sed 's/\^\[\[0m//g' | sed 's/\$$//' | awk 'BEGIN {OFS = ""} {print "  ",$0}'`
+			if [ "$RET0" != "" ]
+			then
+				TOTAL=`printf "$RET0" | wc -l | sed 's/ //g'`
+				printf $C_RED"  $TOTAL failed test(s)"$C_CLEAR
+			else
+				printf $C_GREEN"  All Unit Tests passed"$C_CLEAR
+			fi
+			RET0=`cat -e .mymoulitest | sed 's/\^\[\[[0-9;]*m//g' | sed 's/\^\[\[0m//g' | sed 's/\$$//'`
+			echo "$RET0" > "$RETURNPATH"/.mymoulitest
 		fi
-		RET0=`cat -e .mymoulitest | sed 's/\^\[\[[0-9;]*m//g' | sed 's/\^\[\[0m//g' | sed 's/\$$//'`
-		echo "$RET0" > "$RETURNPATH"/.mymoulitest
+	else
+		printf $C_RED"  'moulitest' is not installed"$C_CLEAR
 	fi
 }
